requirements.txt
fastapi
uvicorn
asyncpg
aiohttp
pyloudnorm
soundfile
numpy
scipy
python-dotenv